fix: proper script termination
The proper way to end termination of a program within a function,
conditional, or loop is to send a signal to the script process ID, trap
the signal, and perform an appropriate exit.

I've created two traps:

- TERM will terminate using "exit 0"
- KILL will terminate using "exit 1"

I've updated every location where `exit` is used to instead send `kill
-s {SIGNAL} $TOP_PID`, where `$TOP_PID` is assigned to the script
process ID and exported at the start of script execution.

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>