feat: Build only latest tags
First, this patch modifies when the workflow is triggered:

- It can trigger for pushes to the master branch (existing behavior).
- It can trigger on release publication (new behavior).
- It can trigger for a manual dispatch (existing behavior, but modified
  as described below).

For pushes to master, we continue as we did before.

For tag releases or manual triggering, the entrypoint now calls on a new
PHP script, `do-we-build-and-if-so-what.php`, at the beginning of
execution. This script does the following:

- It pulls a list of releases, filtering out non-stable releases and
  sorting them via semver in ASC order, and then detects the latest
  stable release.
- If it detects that the GITHUB_REF is a sha1, it echoes the name of the
  latest stable release, returning execution to the entrypoint.
- If it detects that the release is non-stable, it echoes "FALSE" and
  returns execution to the entrypoint, which will then skip the build.
- If it detects that the release is not the latest stable release, it
  echoes "FALSE" and returns execution to the entrypoint, which will then
  skip the build.
- Otherwise, it returns the release reference (its name) back to the
  entrypoint, which then uses it to build docs.

This approach should allow the action to continue to work on existing
repositories, and then allow updating those that are using release
branches to adopt the release-based workflow.

Additionally, it updates the README to indicate the workflow should be
done for releases and manual builds.

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>