feat: Detect if we need to build or not
This patch updates the entrypoint to call a PHP script, `determine_if_we_need_to_build.php`,
at the beginning of execution. This script does the following:

- If the GITHUB_REF is `master`, it echoes "TRUE", to indicate we should build.
- If the GITHUB_REF does not match `^\d+\.\d+\.x`, it indicates NOT to build.
- It then pulls branches, filtering out that do not match the above pattern
- It compares the most recent branch (per semver) with the current
  branch. If they are the same, it indicates we should build
- It then pulls tags, and finds the most recent tag. If the major/minor
  pair match that of the pushed branch, it indicates we should build;
  otherwise, it indicates we should NOT build.

This approach should allow the action to continue to work on existing
repositories, and then allow updating those that are using release
branches to allow any push.

Additionally, it updates the README to indicate the workflow should be
done for pushes to any branches. While the PHP script will check against
any ref, this should help restrict how often it operates.

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>