fix: better detection of different event and ref types
The GITHUB_REF variable holds content in the form
`refs/(tags|heads)/{name}`, so we need to normalize the string before we
attempt to use it for comparisons to branches or versions.

When we do a repository_dispatch, the GITHUB_REF is always the name of
the default branch. Since that changes, it's better if we detect via the
event name.

This patch updates the previous to:

- Detect if we  have a "push" event against the master branch, and, if
  so, `buildDocs('master')`.
- Detect if the event name is NEITHER repository_dispatch NOR release at
  this point; if so, we `skipDocs()`.
- Detect if we have a repository_dispatch event; if so, we
  `buildDocs($latestStableRelease)`.
- At this point, we have a release event; we check to see if the release
  name is the latest; if so, we `buildDocs($latestStableRelease)`,
  otherwise we `skipDocs()`.

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>