feat: prefer a DEPLOY_TOKEN over DOCS_DEPLOY_KEY
In reviewing issues with documentation publication via the docs-build action, the problem appears to be with permissions to checkout and/or push to the source repository running the action.
This is a solved problem with laminas/automatic-releases: you can use git over HTTPS with a privileged GitHub OAuth token (generally in the ORGANIZATION_ADMIN_TOKEN organization-level secret).

This patch modifies the setup of the "remote_repo" variable in a backwards-compatible way.
If a DEPLOY_TOKEN is present in the workflow env (e.g. `DEPLOY_TOKEN: {{ secrets.ORGANIZATION_ADMIN_TOKEN }}`), then this will be used to construct a git+HTTPS URL with the token.
If a DOCS_DEPLOY_KEY is present in the workflow env (current usage), it will set it up as the root user SSH key.
Otherwise, it will fail (same as current behavior).

This allows repositories to opt-in to the new behavior, allowing us to test and refine it, without breaking usage for existing repositories.

While performing these changes, shellcheck flagged some potentially unsafe or erroneous syntax, which I have also fixed (primarily lack of quoting when using variables, and failure to wrap the left side of A && B || C statements in parens).

I have also documented usage updates in the readme.

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>