fix: use x-access-token format
As documented here: https://stackoverflow.com/a/58393457

In place of basic auth details, you can use `x-access-token:${token}`.
Since we saw an issue with usage of the action that detailed "could not read Password for ...", this may be the solution.

Again, tested using this to both pull and push to a private repository, and encountered no issues.

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>