Add abandoned package notification to site pages
Included a notice in both the installation and homepage to inform users when a package is abandoned. This notice is based on the condition that 'abandoned' is present in the lowercase form of the site name. It aims to clearly communicate to users that the package will not receive further updates or development.

Signed-off-by: Frank Brückner <dev@froschdesignstudio.de>